backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
hyperliquid-python-sdk==0.18.0
python-dotenv==1.0.0 